\name{burposte}
\Rdversion{1.1}
\alias{burposte}
\docType{data}
\title{Burposte data}
\description{Burposte data}
\usage{data(burposte)}
\format{
  A data frame with 9521 observations on the following 3 variables.
  \describe{
    \item{\code{ident}}{a numeric vector}
    \item{\code{xinput}}{a numeric vector}
    \item{\code{yprod}}{a numeric vector}
  }
}
\references{
Cazals C., Florens J.-P., Simar L. (2002), Nonparametric frontier estimation: a robust approach, \emph{Journal of Econometrics 106}, 1-25.
}
\examples{
data(burposte)
}
\keyword{datasets}
