\name{spain}
\Rdversion{1.1}
\alias{spain}
\docType{data}
\title{
Spain data
}
\description{
Spain data
}
\usage{data(spain)}
\format{
  A data frame with 61 observations on the following 4 variables.
  \describe{
    \item{\code{y}}{a numeric vector of output}
    \item{\code{x1}}{a numeric vector of input}
    \item{\code{x2}}{a numeric vector of input}
    \item{\code{x3}}{a numeric vector of input}
  }
}


\examples{
data(spain)
}
\keyword{datasets}
